flask_app/requirements.txt
-i https://pypi.org/simple
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
filelock==3.12.4
flask==3.0.0
fonttools==4.43.1
fsspec==2023.10.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jinja2==3.1.2
kiwisolver==1.4.5
markupsafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.1.0
torchvision==0.16
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
werkzeug==3.0.0
zipp==3.17.0
